# docs/openapi.json
{
  "components": {
    "schemas": {
      "ApiResponse_GraphSubgraphView_": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfidenceView"
              },
              {
                "type": "null"
              }
            ]
          },
          "data": {
            "$ref": "#/components/schemas/GraphSubgraphView"
          },
          "envelope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OutputEnvelope"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            },
            "title": "Evidence",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvenanceSummary"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "scope": {
            "$ref": "#/components/schemas/ResponseScope"
          },
          "stage_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Detail"
          },
          "stage_status": {
            "$ref": "#/components/schemas/StageStatusView"
          },
          "standing": {
            "$ref": "#/components/schemas/GraphStandingView"
          },
          "standing_notice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Standing Notice"
          },
          "timing": {
            "$ref": "#/components/schemas/TimingMeta"
          }
        },
        "required": [
          "scope",
          "provenance",
          "standing",
          "stage_status",
          "timing",
          "generated_at",
          "data"
        ],
        "title": "ApiResponse[GraphSubgraphView]",
        "type": "object"
      },
...
      "ApiResponse_Page_EventView__": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfidenceView"
              },
              {
                "type": "null"
              }
            ]
          },
          "data": {
            "$ref": "#/components/schemas/Page_EventView_"
          },
          "envelope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OutputEnvelope"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            },
            "title": "Evidence",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvenanceSummary"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "scope": {
            "$ref": "#/components/schemas/ResponseScope"
          },
          "stage_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Detail"
          },
          "stage_status": {
            "$ref": "#/components/schemas/StageStatusView"
          },
          "standing": {
            "$ref": "#/components/schemas/GraphStandingView"
          },
          "standing_notice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Standing Notice"
          },
          "timing": {
            "$ref": "#/components/schemas/TimingMeta"
          }
        },
        "required": [
          "scope",
          "provenance",
          "standing",
          "stage_status",
          "timing",
          "generated_at",
          "data"
        ],
        "title": "ApiResponse[Page[EventView]]",
        "type": "object"
      },
...
      "ApiResponse_dict_str__Any__": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfidenceView"
              },
              {
                "type": "null"
              }
            ]
          },
          "data": {
            "title": "Data",
            "type": "object"
          },
          "envelope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OutputEnvelope"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            },
            "title": "Evidence",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvenanceSummary"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "scope": {
            "$ref": "#/components/schemas/ResponseScope"
          },
          "stage_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Detail"
          },
          "stage_status": {
            "$ref": "#/components/schemas/StageStatusView"
          },
          "standing": {
            "$ref": "#/components/schemas/GraphStandingView"
          },
          "standing_notice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Standing Notice"
          },
          "timing": {
            "$ref": "#/components/schemas/TimingMeta"
          }
        },
        "required": [
          "scope",
          "provenance",
          "standing",
          "stage_status",
          "timing",
          "generated_at",
          "data"
        ],
        "title": "ApiResponse[dict[str, Any]]",
        "type": "object"
      },
...
      "ApiResponse_tuple_RunView__________": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfidenceView"
              },
              {
                "type": "null"
              }
            ]
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/RunView"
            },
            "title": "Data",
            "type": "array"
          },
          "envelope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OutputEnvelope"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            },
            "title": "Evidence",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvenanceSummary"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "scope": {
            "$ref": "#/components/schemas/ResponseScope"
          },
          "stage_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Detail"
          },
          "stage_status": {
            "$ref": "#/components/schemas/StageStatusView"
          },
          "standing": {
            "$ref": "#/components/schemas/GraphStandingView"
          },
          "standing_notice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Standing Notice"
          },
          "timing": {
            "$ref": "#/components/schemas/TimingMeta"
          }
        },
        "required": [
          "scope",
          "provenance",
          "standing",
          "stage_status",
          "timing",
          "generated_at",
          "data"
        ],
        "title": "ApiResponse[tuple[RunView, ...]]",
        "type": "object"
      },
...
      "ConfidenceView": {
        "additionalProperties": false,
        "description": "A confidence number that cannot exist without its decomposition.\n\n`components` carries `min_length=1`, so **the API is structurally incapable of\nreturning a bare float wearing a confidence's name**. LAW-EVIDENCE says a bare float is\na defect; this makes it an unconstructable one rather than something a test has to go\nlooking for. It is the same move ADR-0075 made for a recommendation's evidence, in the\nplace a client actually reads.\n\n`scalar` is derived and names the `aggregation` that produced it, exactly as\n`ConfidenceVector` does. A consumer that reads `scalar` and ignores `components` has\nbeen given everything it needs to know better.",
        "properties": {
          "aggregation": {
            "minLength": 1,
            "title": "Aggregation",
            "type": "string"
          },
          "components": {
            "items": {
              "$ref": "#/components/schemas/ConfidenceComponentView"
            },
            "minItems": 1,
            "title": "Components",
            "type": "array"
          },
          "provenance_class": {
            "$ref": "#/components/schemas/ProvenanceClass"
          },
          "scalar": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Scalar",
            "type": "number"
          }
        },
        "required": [
          "scalar",
          "aggregation",
          "provenance_class",
          "components"
        ],
        "title": "ConfidenceView",
        "type": "object"
      },
...
      "GraphStandingView": {
        "description": "Which graph a traversal walked (ADR-0059), mirrored for the presentation layer.\n\nA mirror rather than a re-export because `GraphStanding` lives in `causal_engine` (L6)\nand the API may not import it. The two are kept in step by\n`tests/api/test_standing_mirror_is_total.py`, which enumerates both enums: a member\nadded to one and not the other fails, so the mirror cannot silently narrow.\n\n`UNPROMOTED_DIAGNOSTIC` is carried into responses ON PURPOSE. It marks output the\nengine has disowned -- walked over module 10's scored graph before promotion -- and\nADR-0059 defends it with a required field, one construction site, a validator refusing\n`INFERRED`, and a fixed notice. OQ-026 records the residual risk honestly: none of\nthose four mechanisms survives a screenshot. The API adds the fifth it can: the notice\nis a REQUIRED field on any response carrying this standing, not an optional annotation.",
        "enum": [
          "STATED",
          "UNPROMOTED_DIAGNOSTIC"
        ],
        "title": "GraphStandingView",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StageStatusView": {
        "description": "Whether the pipeline stage behind a payload actually produced it.\n\n`NOT_RUNNABLE` is what an endpoint returns when its owning module has not been built --\nmodules 7, 8 and 15 today. It is a 200 with a full envelope and an explicit statement,\nnever an empty success: `docs/architecture.md` \u00a72 requires that a partially computed\nrun \"returns what exists with the stage stated, never a silently truncated graph\", and\nan empty list with no explanation is exactly the silent truncation that forbids.",
        "enum": [
          "COMPLETE",
          "PARTIAL",
          "NOT_RUNNABLE"
        ],
        "title": "StageStatusView",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "A domain-agnostic causal intelligence engine. Every response carries its run, its provenance, its evidence references and its confidence decomposition; no endpoint returns a confidence without its components.",
    "title": "CausaLog",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/runs": {
      "get": {
        "description": "List every run materialized in this process, in canonical sequence.\n\nCATALOG scope: this body spans many runs, so there is no single `OutputEnvelope` that\ndescribes it. Each ITEM carries the five inputs that identify its run; synthesizing an\nenvelope for the listing would be fabricating version fields that describe nothing.",
        "operationId": "list_runs_v1_runs_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_tuple_RunView__________"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/v1/runs/{run_id}/events": {
      "get": {
        "description": "Return a page of a run's events, in `CONVENTIONS.md` \u00a711's canonical sequence.\n\nOne extra item is requested and then dropped. That is how `has_more` is answered\nwithout a second query and without a count: asking for `size + 1` and seeing whether it\narrived distinguishes \"this is the last page\" from \"the next page is empty\", which a\ncursor alone cannot.",
        "operationId": "read_events_v1_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Page_EventView__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Events",
        "tags": [
          "analysis"
        ]
      }
    },
    "/v1/runs/{run_id}/graph": {
      "get": {
        "description": "Extract a bounded subgraph, saying what was left out and why it may be empty.",
        "operationId": "read_graph_v1_runs__run_id__graph_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "seed_event_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Seed Event Id"
            }
          },
          {
            "in": "query",
            "name": "depth",
            "required": false,
            "schema": {
              "default": 2,
              "maximum": 10,
              "minimum": 0,
              "title": "Depth",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "standing",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/GraphStandingView",
              "default": "STATED"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_GraphSubgraphView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Graph",
        "tags": [
          "analysis"
        ]
      }
    },
...
    "/v1/runs/{run_id}/propagation/{seed_event_id}": {
      "get": {
        "description": "Measure one event's consequences, attributed to the node set (ADR-0061).\n\nThe payload is the canonical body of a `draft` artifact rather than a hand-written\nview: `PropagationReport` is `draft` in `CONTEXT.md` \u00a76, and mirroring it into a frozen\nwire shape would freeze it by assertion. `docs/api.md` marks this body as `draft`.",
        "operationId": "read_propagation_v1_runs__run_id__propagation__seed_event_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "seed_event_id",
            "required": true,
            "schema": {
              "title": "Seed Event Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "standing",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/GraphStandingView",
              "default": "STATED"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__Any__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Propagation",
        "tags": [
          "analysis"
        ]
      }
    },
    "/v1/runs/{run_id}/root-causes/{outcome_event_id}": {
      "get": {
        "description": "Rank the causes of one outcome as ADR-0008's four never-merged views.",
        "operationId": "read_root_causes_v1_runs__run_id__root_causes__outcome_event_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "outcome_event_id",
            "required": true,
            "schema": {
              "title": "Outcome Event Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "standing",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/GraphStandingView",
              "default": "STATED"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RootCauseView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Root Causes",
        "tags": [
          "analysis"
        ]
      }
    },
...
    "/v1/runs/{run_id}/timelines/{process_instance_id}": {
      "get": {
        "description": "Return one process instance's timeline, with gaps marked rather than interpolated.",
        "operationId": "read_timeline_v1_runs__run_id__timelines__process_instance_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "process_instance_id",
            "required": true,
            "schema": {
              "title": "Process Instance Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Union_TimelineView__NoneType__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Timeline",
        "tags": [
          "analysis"
        ]
      }
    }
  }
}